image_upload_pipeline/image_pipeline-main/requirements.txt
boto3
datasets
huggingface_hub
Pillow
